+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved `NSSInput` and `NSSOutput` to subpackage `nss`
 - Renamed `APEIRecipeMapper` to `ARecipeTypeMapper`
 - A whole codebase cleanup
+- Move lots of functions from `ARecipeTypeMapper` to `Utils`
 
 ## [4.1.2]
 ### Added